fix an over-release bug on images. More migration to Xcode assets. Attempt to make the app 64-bit safe. It'll take a lot of work.

+++ win/nh3d/NH3DPanel.m
@@ -3,7 +3,7 @@
 @implementation NH3DPanel
 
 - (id)initWithContentRect:(NSRect)contentRect 
-				styleMask:(unsigned int)aStyle 
+				styleMask:(NSUInteger)aStyle
 				  backing:(NSBackingStoreType)bufferingType 
 					defer:(BOOL)flag
 {
